use applicative instead of monad
